DataBlockHashIndex: explicit type conversion in Initialize() param

+++ table/data_block_hash_index.cc
@@ -74,7 +74,8 @@ void DataBlockHashIndex::Initialize(const char* data, uint16_t size,
   num_buckets_ = DecodeFixed16(data + size - sizeof(uint16_t));
   assert(num_buckets_ > 0);
   assert(size > num_buckets_ * sizeof(uint8_t));
-  *map_offset = size - sizeof(uint16_t) - num_buckets_ * sizeof(uint8_t);
+  *map_offset = static_cast<uint16_t>(size - sizeof(uint16_t) -
+                                      num_buckets_ * sizeof(uint8_t));
 }
 
 uint8_t DataBlockHashIndex::Seek(const char* data, uint16_t map_offset,
